[Refactor] Move storage workflow tasks into DataWorkflows
