+++ be/src/agent/agent_task.cpp
@@ -25,6 +25,8 @@
 #include "common/config_agent_fwd.h"
 #include "common/status.h"
 #include "common/system/backend_options.h"
+#include "data_workflows/migration/engine_storage_migration_task.h"
+#include "data_workflows/schema_change/engine_alter_tablet_task.h"
 #include "gutil/strings/join.h"
 #include "io/io_profiler.h"
 #include "runtime/current_thread.h"
@@ -38,12 +40,10 @@
 #include "storage/snapshot_manager.h"
 #include "storage/storage_env.h"
 #include "storage/tablet_manager.h"
-#include "storage/task/engine_alter_tablet_task.h"
 #include "storage/task/engine_checksum_task.h"
 #include "storage/task/engine_clone_task.h"
 #include "storage/task/engine_compaction_control_task.h"
 #include "storage/task/engine_manual_compaction_task.h"
-#include "storage/task/engine_storage_migration_task.h"
 #include "storage/txn_manager.h"
 #include "storage/update_manager.h"
