+++ be/src/data_workflows/load/engine_batch_load_task.cpp
@@ -42,10 +42,10 @@
 
 #include "base/utility/defer_op.h"
 #include "base/utility/pretty_printer.h"
+#include "data_workflows/load/lake_spark_load_handler.h"
 #include "data_workflows/load/push_handler.h"
 #include "runtime/current_thread.h"
 #include "runtime/exec_env.h"
-#include "storage/lake/spark_load.h"
 #include "storage/lake/tablet_manager.h"
 #include "storage/lake/versioned_tablet.h"
 #include "storage/olap_common.h"
