+++ be/src/data_workflows/load/push_handler.h
@@ -16,10 +16,10 @@
 
 #include <vector>
 
+#include "data_workflows/load/push_broker_reader.h"
 #include "exec/file_scanner/file_scanner.h"
 #include "runtime/descriptors.h"
 #include "storage/olap_common.h"
-#include "storage/push_utils.h"
 #include "storage/rowset/rowset.h"
 #include "storage/tablet.h"
