+++ build-support/exec_env_singleton_allowlist.txt
@@ -15,7 +15,7 @@ src/service/service_be/arrow_flight_sql_service.cpp
 src/service/service_be/lake_service.cpp
 src/service/service_be/starrocks_be.cpp
 src/storage/load_chunk_spiller.cpp
-src/storage/push_utils.cpp
+src/data_workflows/load/push_broker_reader.cpp
 src/storage/rowset/json_column_iterator.cpp
 src/storage/schema_change_utils.cpp
 src/udf/java/utils.cpp
